+++ src/RendererManager.cpp
@@ -18,7 +18,7 @@ void	RendererManager::initGraphics(const std::string &name)
   _gc = _window->get_gc();
 }
 
-void	RendererManager::update(GameState &state, int elapsedTime)
+void	RendererManager::update(GameState &state, double elapsedTime)
 {
   if ((state.getPaused() & GameState::NODRAW) == GameState::NODRAW)
     return ;
